- Moved get_default_scan_params and get_dda_scan_params out from the environment into Common.py. - Polarity now is an optional parameter to these two methods (defaults to POSITIVE). - Had to move the ScanParameters and Precursor classes as well to Common.py due to circular imports (Common.py imports from MassSpec.py which imports from Common.py)
