+++ vimms/Environment.py
@@ -7,9 +7,9 @@ from vimms.Common import DEFAULT_MS1_SCAN_WINDOW, DEFAULT_ISOLATION_WIDTH, DEFAU
     DEFAULT_MS1_ORBITRAP_RESOLUTION, DEFAULT_MS1_AGC_TARGET, DEFAULT_MS1_MAXIT, DEFAULT_MS2_COLLISION_ENERGY, \
     DEFAULT_MS2_ORBITRAP_RESOLUTION, DEFAULT_MS2_AGC_TARGET, DEFAULT_MS2_MAXIT, POSITIVE, DEFAULT_MSN_SCAN_WINDOW, \
     DEFAULT_MS1_MASS_ANALYSER, DEFAULT_MS1_ACTIVATION_TYPE, DEFAULT_MS1_ISOLATION_MODE, DEFAULT_MS2_MASS_ANALYSER, \
-    DEFAULT_MS2_ISOLATION_MODE, DEFAULT_SOURCE_CID_ENERGY
+    DEFAULT_MS2_ISOLATION_MODE, DEFAULT_SOURCE_CID_ENERGY, ScanParameters, Precursor
 from vimms.Controller import TopNController, PurityController
-from vimms.MassSpec import ScanParameters, IndependentMassSpectrometer, Precursor
+from vimms.MassSpec import IndependentMassSpectrometer
 from vimms.MzmlWriter import MzmlWriter
 
 
@@ -185,103 +185,6 @@ class Environment(object):
 
     def get_initial_scan_params(self):
         return self.controller.get_initial_scan_params()
-
-    def get_default_scan_params(self, agc_target=DEFAULT_MS1_AGC_TARGET, max_it=DEFAULT_MS1_MAXIT,
-                                collision_energy=DEFAULT_MS1_COLLISION_ENERGY,
-                                source_cid_energy=DEFAULT_SOURCE_CID_ENERGY,
-                                orbitrap_resolution=DEFAULT_MS1_ORBITRAP_RESOLUTION,
-                                default_ms1_scan_window=DEFAULT_MS1_SCAN_WINDOW,
-                                mass_analyser=DEFAULT_MS1_MASS_ANALYSER,
-                                activation_type=DEFAULT_MS1_ACTIVATION_TYPE,
-                                isolation_mode=DEFAULT_MS1_ISOLATION_MODE,
-                                metadata=None):
-        """
-        Gets the default method scan parameters. Now it's set to do MS1 scan only.
-        :return: the default scan parameters
-        """
-        default_scan_params = ScanParameters()
-        default_scan_params.set(ScanParameters.MS_LEVEL, 1)
-        default_scan_params.set(ScanParameters.ISOLATION_WINDOWS, [[default_ms1_scan_window]])
-        default_scan_params.set(ScanParameters.ISOLATION_WIDTH, DEFAULT_ISOLATION_WIDTH)
-        default_scan_params.set(ScanParameters.COLLISION_ENERGY, collision_energy)
-        default_scan_params.set(ScanParameters.ORBITRAP_RESOLUTION, orbitrap_resolution)
-        default_scan_params.set(ScanParameters.ACTIVATION_TYPE, activation_type)
-        default_scan_params.set(ScanParameters.MASS_ANALYSER, mass_analyser)
-        default_scan_params.set(ScanParameters.ISOLATION_MODE, isolation_mode)
-        default_scan_params.set(ScanParameters.AGC_TARGET, agc_target)
-        default_scan_params.set(ScanParameters.MAX_IT, max_it)
-        default_scan_params.set(ScanParameters.SOURCE_CID_ENERGY, source_cid_energy)
-        default_scan_params.set(ScanParameters.POLARITY, self.mass_spec.ionisation_mode)
-        default_scan_params.set(ScanParameters.FIRST_MASS, default_ms1_scan_window[0])
-        default_scan_params.set(ScanParameters.LAST_MASS, default_ms1_scan_window[1])
-        default_scan_params.set(ScanParameters.METADATA, metadata)
-        return default_scan_params
-
-    def get_dda_scan_param(self, mz, intensity, precursor_scan_id, isolation_width, mz_tol, rt_tol,
-                           agc_target=DEFAULT_MS2_AGC_TARGET, max_it=DEFAULT_MS2_MAXIT,
-                           collision_energy=DEFAULT_MS2_COLLISION_ENERGY,
-                           source_cid_energy=DEFAULT_SOURCE_CID_ENERGY,
-                           orbitrap_resolution=DEFAULT_MS2_ORBITRAP_RESOLUTION,
-                           mass_analyser=DEFAULT_MS2_MASS_ANALYSER,
-                           activation_type=DEFAULT_MS1_ACTIVATION_TYPE,
-                           isolation_mode=DEFAULT_MS2_ISOLATION_MODE,
-                           metadata=None):
-        dda_scan_params = ScanParameters()
-        dda_scan_params.set(ScanParameters.MS_LEVEL, 2)
-
-        assert isinstance(mz, list) == isinstance(intensity, list)
-
-        # create precursor object, assume it's all singly charged
-        precursor_charge = +1 if (self.mass_spec.ionisation_mode == POSITIVE) else -1
-        if type(mz) == list:
-            precursor_list = []
-            for i, m in enumerate(mz):
-                precursor_list.append(Precursor(precursor_mz=m, precursor_intensity=intensity[i],
-                                                precursor_charge=precursor_charge, precursor_scan_id=precursor_scan_id))
-            dda_scan_params.set(ScanParameters.PRECURSOR_MZ, precursor_list)
-
-            if type(isolation_width) == list:
-                assert len(isolation_width) == len(precursor_list)
-            else:
-                isolation_width = [isolation_width for m in mz]
-            dda_scan_params.set(ScanParameters.ISOLATION_WIDTH, isolation_width)
-
-        else:
-            precursor = Precursor(precursor_mz=mz, precursor_intensity=intensity,
-                                  precursor_charge=precursor_charge, precursor_scan_id=precursor_scan_id)
-            precursor_list = [precursor]
-            dda_scan_params.set(ScanParameters.PRECURSOR_MZ, precursor_list)
-
-            # set the full-width isolation width, in Da
-            # if mz is not a list, neither should isolation_width be
-            assert not isinstance(isolation_width, list)
-            isolation_width = [isolation_width]
-            dda_scan_params.set(ScanParameters.ISOLATION_WIDTH, isolation_width)
-
-        # define dynamic exclusion parameters
-        dda_scan_params.set(ScanParameters.DYNAMIC_EXCLUSION_MZ_TOL, mz_tol)
-        dda_scan_params.set(ScanParameters.DYNAMIC_EXCLUSION_RT_TOL, rt_tol)
-
-        # define other fragmentation parameters
-        dda_scan_params.set(ScanParameters.COLLISION_ENERGY, collision_energy)
-        dda_scan_params.set(ScanParameters.ORBITRAP_RESOLUTION, orbitrap_resolution)
-        dda_scan_params.set(ScanParameters.ACTIVATION_TYPE, activation_type)
-        dda_scan_params.set(ScanParameters.MASS_ANALYSER, mass_analyser)
-        dda_scan_params.set(ScanParameters.ISOLATION_MODE, isolation_mode)
-        dda_scan_params.set(ScanParameters.AGC_TARGET, agc_target)
-        dda_scan_params.set(ScanParameters.MAX_IT, max_it)
-        dda_scan_params.set(ScanParameters.SOURCE_CID_ENERGY, source_cid_energy)
-        dda_scan_params.set(ScanParameters.POLARITY, self.mass_spec.ionisation_mode)
-        dda_scan_params.set(ScanParameters.FIRST_MASS, DEFAULT_MSN_SCAN_WINDOW[0])
-        dda_scan_params.set(ScanParameters.METADATA, metadata)
-
-        # dynamically scale the upper mass
-        charge = 1
-        wiggle_room = 1.1
-        max_precursor_mz = max([(p.precursor_mz + isol/2) for (p, isol) in zip(precursor_list, isolation_width)])
-        last_mass = max_precursor_mz * charge * wiggle_room
-        dda_scan_params.set(ScanParameters.LAST_MASS, last_mass)
-        return dda_scan_params
 
     def _get_N_DEW(self, time):
         """
