+++ vimms/PythonMzmine.py
@@ -6,8 +6,7 @@ import pandas as pd
 from loguru import logger
 
 from vimms.Chemicals import UnknownChemical
-from vimms.Common import PROTON_MASS
-from vimms.MassSpec import ScanParameters
+from vimms.Common import PROTON_MASS, ScanParameters
 from vimms.PlotsForPaper import get_chem_frag_counts, update_matched_status, compute_pref_rec_f1, get_frag_events
 from vimms.SpectralUtils import get_scans
