+++ vimms/SpectralUtils.py
@@ -4,8 +4,8 @@ import pandas as pd
 import pymzml
 from loguru import logger
 
-from vimms.Common import get_rt
-from vimms.MassSpec import Scan, ScanParameters, Precursor
+from vimms.Common import get_rt, ScanParameters, Precursor
+from vimms.MassSpec import Scan
 from vimms.Roi import make_roi
 from vimms.Chemicals import RoiToChemicalCreator
